Add initial `dropdown` directive implementation
