+++ sphinx_panels/css/bs-cards.css
@@ -204,3 +204,92 @@
     .text-center {
     text-align:center!important
     }
+    .font-weight-light {
+    font-weight:300!important
+    }
+    .font-weight-lighter {
+    font-weight:lighter!important
+    }
+    .font-weight-normal {
+    font-weight:400!important
+    }
+    .font-weight-bold {
+    font-weight:700!important
+    }
+    .font-weight-bolder {
+    font-weight:bolder!important
+    }
+    .font-italic {
+    font-style:italic!important
+    }
+    .text-white {
+        color:#fff!important
+    }
+    .text-primary {
+    color:#007bff!important
+    }
+    a.text-primary:focus,
+    a.text-primary:hover {
+    color:#0056b3!important
+    }
+    .text-secondary {
+    color:#6c757d!important
+    }
+    a.text-secondary:focus,
+    a.text-secondary:hover {
+    color:#494f54!important
+    }
+    .text-success {
+    color:#28a745!important
+    }
+    a.text-success:focus,
+    a.text-success:hover {
+    color:#19692c!important
+    }
+    .text-info {
+    color:#17a2b8!important
+    }
+    a.text-info:focus,
+    a.text-info:hover {
+    color:#0f6674!important
+    }
+    .text-warning {
+    color:#ffc107!important
+    }
+    a.text-warning:focus,
+    a.text-warning:hover {
+    color:#ba8b00!important
+    }
+    .text-danger {
+    color:#dc3545!important
+    }
+    a.text-danger:focus,
+    a.text-danger:hover {
+    color:#a71d2a!important
+    }
+    .text-light {
+    color:#f8f9fa!important
+    }
+    a.text-light:focus,
+    a.text-light:hover {
+    color:#cbd3da!important
+    }
+    .text-dark {
+    color:#343a40!important
+    }
+    a.text-dark:focus,
+    a.text-dark:hover {
+    color:#121416!important
+    }
+    .text-body {
+    color:#212529!important
+    }
+    .text-muted {
+    color:#6c757d!important
+    }
+    .text-black-50 {
+    color:rgba(0,0,0,.5)!important
+    }
+    .text-white-50 {
+    color:rgba(255,255,255,.5)!important
+    }
